+++ src/main/java/com/olivia/peanut/aps/model/ApsMachineWorkstation.java
@@ -40,7 +40,7 @@ public class ApsMachineWorkstation extends BaseEntity<ApsMachineWorkstation> {
    *  最大功率
    */
   @TableField("max_power")
-  private BigDecimal maxPower;
+  private Integer maxPower;
   /***
    *  工厂ID
    */
