+++ src/main/java/com/olivia/peanut/aps/model/ApsProduceProcessItem.java
@@ -64,7 +64,7 @@ public class ApsProduceProcessItem extends BaseEntity<ApsProduceProcessItem> {
    *  最大功率
    */
   @TableField("max_power")
-  private BigDecimal maxPower;
+  private Integer maxPower;
   /***
    *  工厂ID
    */
